moved to_json to fusionbox.utils
