+++ fusionbox/templatetags/fusionbox_tags.py
@@ -4,7 +4,6 @@ from decimal import Decimal, ROUND_HALF_UP, InvalidOperation
 import locale
 locale.setlocale(locale.LC_ALL, 'en_US.UTF-8')
 
-import datetime
 import re
 import warnings
 
@@ -22,11 +21,11 @@ try:
 except ImportError:
     pass
 
+from fusionbox.utils import to_json
 from django import template
 from django.conf import settings
 
 from BeautifulSoup import BeautifulSoup
-from django.utils import simplejson
 from django.utils.safestring import mark_safe
 from django.contrib.humanize.templatetags.humanize import intcomma
 from django.core.exceptions import ImproperlyConfigured
@@ -202,41 +201,6 @@ def attr(obj, arg1):
     return obj
 
 
-def more_json(obj):
-    """
-    Allows decimals and objects with `to_json` methods to be serialized.
-    """
-    if isinstance(obj, Decimal):
-        return float(obj)
-    if isinstance(obj, datetime.datetime):
-        return {
-            '__class__': 'datetime.datetime',
-            'year': obj.year,
-            'month': obj.month,
-            'day': obj.day,
-            'hour': obj.hour,
-            'minute': obj.minute,
-            'second': obj.second,
-            }
-    if isinstance(obj, datetime.time):
-        return {
-            '__class__': 'datetime.time',
-            'hour': obj.hour,
-            'minute': obj.minute,
-            'second': obj.second,
-            }
-    if isinstance(obj, datetime.date):
-        return {
-            '__class__': 'datetime.date',
-            'year': obj.year,
-            'month': obj.month,
-            'day': obj.day,
-            }
-    if hasattr(obj, 'to_json'):
-        return obj.to_json()
-    raise TypeError("%r is not JSON serializable" % (obj,))
-
-
 @register.filter
 def json(a):
     """
@@ -249,15 +213,7 @@ def json(a):
     If the output needs to be put in an attribute, entitize the output of this
     filter.
     """
-    json_str = simplejson.dumps(a, default=more_json)
-
-    # Escape all the XML/HTML special characters.
-    escapes = ['<', '>', '&']
-    for c in escapes:
-        json_str = json_str.replace(c, r'\u%04x' % ord(c))
-
-    # now it's safe to use mark_safe
-    return mark_safe(json_str)
+    return mark_safe(to_json(a))
 json.is_safe = True
